requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
paramiko>=3.3.0
pyyaml>=6.0
requests>=2.31.0
psutil>=5.9.0
